ilcd: use plain ints as exchange IDs; ...
add ../external_docs/ prefix for external files in sources export
(compatibility with ILCD and soda4LCA)

+++ olca-ilcd/src/main/java/org/openlca/ilcd/processes/AllocationFactor.java
@@ -2,8 +2,6 @@
 package org.openlca.ilcd.processes;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
-import java.math.BigInteger;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
@@ -17,9 +15,9 @@ public class AllocationFactor implements Serializable {
 	private final static long serialVersionUID = 1L;
 
 	@XmlAttribute(name = "internalReferenceToCoProduct")
-	public BigInteger referenceToCoProduct;
+	public int referenceToCoProduct;
 
 	@XmlAttribute(name = "allocatedFraction")
-	public BigDecimal allocatedFraction;
+	public double allocatedFraction;
 
 }
